+++ variables.tf
@@ -45,3 +45,69 @@ variable "extra_tags" {
   type        = map(string)
   default     = {}
 }
+
+variable "deny_all_inbound" {
+  description = "True to deny all inbound traffic by default"
+  type        = bool
+  default     = false
+}
+
+variable "http_inbound_allowed" {
+  description = "True to allow inbound HTTP traffic"
+  type        = bool
+  default     = false
+}
+
+variable "allowed_http_cidr_list" {
+  description = "List of CIDR allowed for inbound HTTP traffic"
+  type        = list(string)
+  default     = []
+}
+
+variable "https_inbound_allowed" {
+  description = "True to allow inbound HTTPS traffic"
+  type        = bool
+  default     = false
+}
+
+variable "allowed_https_cidr_list" {
+  description = "List of CIDR allowed for inbound HTTPS traffic"
+  type        = list(string)
+  default     = []
+}
+
+variable "ssh_inbound_allowed" {
+  description = "True to allow inbound SSH traffic"
+  type        = bool
+  default     = false
+}
+
+variable "allowed_ssh_cidr_list" {
+  description = "List of CIDR allowed for inbound SSH traffic"
+  type        = list(string)
+  default     = []
+}
+
+variable "rdp_inbound_allowed" {
+  description = "True to allow inbound RDP traffic"
+  type        = bool
+  default     = false
+}
+
+variable "allowed_rdp_cidr_list" {
+  description = "List of CIDR allowed for inbound RDP traffic"
+  type        = list(string)
+  default     = []
+}
+
+variable "winrm_inbound_allowed" {
+  description = "True to allow inbound WinRM traffic"
+  type        = bool
+  default     = false
+}
+
+variable "allowed_winrm_cidr_list" {
+  description = "List of CIDR allowed for inbound WinRM traffic"
+  type        = list(string)
+  default     = []
+}